requirement.txt
tensorflow
opencv-python
numpy
fastapi
uvicorn
pyyaml
Pillow
python-multipart